refactor: feature/domain 패키지로 분류
