+++ src/main/java/baseball/Application.java
@@ -1,5 +1,8 @@
 package baseball;
 
+import baseball.domain.Game;
+import baseball.feature.GameOver;
+
 public class Application {
 
     public static void main(String[] args) {
