+++ src/main/java/baseball/feature/GameOver.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.feature;
 
 import camp.nextstep.edu.missionutils.Console;
