+++ src/main/java/baseball/feature/GamePrint.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.feature;
 
 public class GamePrint {
